Add files via upload
The methodology encompasses several key components, with a focus on utilizing EasyOCR for extracting text from images. EasyOCR is a robust optical character recognition (OCR) tool that can effectively analyze images and convert text into a digital format. By integrating EasyOCR into the Artificial Eye, the system can accurately interpret text from captured images, enabling visually impaired individuals to access information in various formats.
OCR is formerly known as Optical Character Recognition which is revolutionary for the digital world nowadays. OCR is actually a complete process under which the images/documents which are present in a digital world are processed and from the text are being processed out as normal editable text.
OCR is a technology that enables you to convert different types of documents, such as scanned paper documents, PDF files, or images captured by a digital camera into editable and searchable data. 
EasyOCR is actually a python package that holds PyTorch as a backend handler. It detects the text from images but in my reference, while using it I found that it is the most straightforward way to detect text from images also when high end deep learning library(PyTorch) is supporting it in the backend which makes it accuracy more credible.